Added Marker Symbols
Felt markers were hard to distinguish based on colors only - added possibility to show symbols for each marker type: Center, Thirds, Mouse position, Custom markers and the newly added Golden Ratio marker. Symbols hard coded so far but carefully selected.

Also added possibility to turn on or off the symbol markers, with default off to make it work like before if nothing done. The settings is placed in settings dialog, advanced tab. German translation added but my German is rather rusty so it was a qualified guess.

When adding symbols in addition to drawing marker position, it became apparent that StringFormatFlags.DirectionRightToLeft is intended for right-to-left languages and not right align text in general. Position only no problem but with symbol added the order of symbol+text changed depending on... I guess the category of language the symbol (unicode) belongs to. Changed to StringAlignment.Far which is the intended format for right alignment - if I understood it correctly. Just a detail, but it took sume time to understand why some symbols came before the marker position string and some after - but only on screen, not as the string in code where they were relatively positioned as expected.... :)